+++ src/App.tsx
@@ -2,9 +2,9 @@ import { RouterProvider } from 'react-router-dom';
 import './App.css';
 import router from './router/router';
 import { Provider } from 'react-redux';
-import { store } from './store/store';
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { store } from './store/store';
 
 function App() {
   return (
