+++ src/modules/Appbar/index.ts
@@ -1,0 +1,2 @@
+export { AppBar } from './components/AppBar/AppBar';
+export { searchValueReducer, setSearchValue } from './store/searchValueSlice';
