+++ src/modules/ImageGallery/components/ImageGallery/ImageGallery.tsx
@@ -2,7 +2,6 @@ import { LinearProgress, useMediaQuery } from '@mui/material';
 import ImageList from '@mui/material/ImageList';
 import { User } from 'firebase/auth';
 import { useAppDispatch, useAppSelector } from '../../../../store/store';
-import { setEditImageData } from '../../store/editImageSlice';
 import { useNavigate } from 'react-router-dom';
 import { EDITOR_ROUTE } from '../../../../constants/routes';
 import ImageItem from '../ImageItem/ImageItem';
@@ -11,6 +10,7 @@ import useSearchByValue from '../../hooks/useSearchByValue';
 import { toast } from 'react-toastify';
 import Modal from '../../../../ui/Modal/Modal';
 import { deleteDocument } from '../../api/deleteImageById';
+import { setEditImageData } from '../../store/editImageSlice';
 
 interface ImageGalleryPropsType {
   user: User | null | undefined;
