+++ src/modules/ToolsPanel/components/ToolPanels/ToolPanels.tsx
@@ -1,0 +1,15 @@
+import { useMediaQuery } from '@mui/material';
+
+import { BrushPanel } from '../BrushPanel/BrushPanel';
+import { ShapesPanel } from '../ShapesPanel/ShapesPanel';
+
+export function ToolPanels() {
+  const isLaptop = useMediaQuery('(min-width:1100px)');
+
+  return (
+    <>
+      <ShapesPanel />
+      <BrushPanel isLaptop={isLaptop} />
+    </>
+  );
+}
