+++ src/modules/ToolsPanel/constants/penWidth.ts
@@ -1,0 +1,22 @@
+export const penWidth = [
+  {
+    value: 1,
+    label: '1px',
+  },
+  {
+    value: 3,
+    label: '3px',
+  },
+  {
+    value: 6,
+    label: '6px',
+  },
+  {
+    value: 9,
+    label: '9px',
+  },
+  {
+    value: 12,
+    label: '12px',
+  },
+];
