+++ src/modules/ToolsPanel/index.ts
@@ -1,0 +1,1 @@
+export { ToolPanels } from './components/ToolPanels/ToolPanels';
