+++ src/modules/ToolsPanel/store/penSizeSlice.ts
