+++ src/modules/ToolsPanel/store/toolColorSlice.ts
