+++ src/modules/ToolsPanel/store/toolSlice.ts
@@ -1,5 +1,5 @@
 import { PayloadAction, createSlice } from '@reduxjs/toolkit';
-import { Tools } from '../constants/tools';
+import { Tools } from '../../../constants/tools';
 
 const initialState = Tools.brush;
