+++ src/pages/EditorPage/EditorPage.tsx
@@ -1,8 +1,8 @@
 import { useAuthState } from 'react-firebase-hooks/auth';
-import ToolsPanel from '../../components/ToolsPanel/ToolsPanel';
 import { auth } from '../../api/config';
 import Spinner from '../../ui/Spinner/Spinner';
 import { Canvas } from '../../modules/Canvas';
+import { ToolPanels } from '../../modules/ToolsPanel';
 
 function EditorPage() {
   const [user, loading] = useAuthState(auth);
@@ -14,7 +14,7 @@ function EditorPage() {
   return (
     <div>
       <Canvas user={user} />
-      <ToolsPanel />
+      <ToolPanels />
     </div>
   );
 }
